// src/app/models/recurso.model.ts
// Generated by https://quicktype.io

export interface RecursoResponse {
  id?:                number;
  tipoRecursoId:     number;
  nombre:            string;
  capacidad:         number | null;
  stock:             number | null;
  ubicacion:         string;
  imagenUrl:         string;
  estado:            Estado;
  fechaCreacion:     Date;
  fechaModificacion?: null;
  historialUsos?:     any[];
  reservas?:          any[];
  tipoRecurso?:       null;
}

export enum Estado {
  Disponible = "Disponible",
  Mantenimiento = "Mantenimiento",
}